+++ services/identity/prisma/migrations/20260924130844_add_email_verification/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE "accounts" ADD COLUMN     "email_verification_expires_at" TIMESTAMP(3),
+ADD COLUMN     "email_verification_token_hash" TEXT;